Solr data directory can now be a symlink.
Previously, the solr:rebuild rake task would use FileUtils.rm_rf to
remove the solr data directory.  This caused issues when the
SolrPowered.data_dir returned the path to a symbolic link.  The link
was removed and then a directory was put in its place.

The new approach removes the contents of the data dir instead of
the directory itself.